requirements.txt
numpy
networkx
python-igraph
pyyaml
pytz
matplotlib
tensorflow==1.15.0